gh-1 removed the old diagram.py
